Temporarily remove support for buildnumber fetch.
This change is so the build will pass with git clones. I'm looking at
patching buildnumber-maven-plugin to support git. wsmoak and pgier on
irc.codehaus.org#maven said that if
http://github.com/astubbs/buildnumber-maven-plugin/commit/a7f99ffb58c2fc087614536bd5792b176fb5ad47
works, they'll apply the patch to
http://mojo.codehaus.org/buildnumber-maven-plugin/source-repository.html

See also:
http://stackoverflow.com/questions/2685069/maven-buildnumber-plugin-git
http://jira.codehaus.org/browse/MOJO-1199

And, perhaps a last resort:
http://www.chrissearle.org/blog/technical/using_maven_exec_plugin_store_current_git_sha_build

This commit is a port of e7b07d3af01cd6b24def92e65e4c48321410d2b6 in head.